Changed graphics structure and added some GameItems
Graphics now operates through a game and a room. The Game is the JFrame
which holds a Room. The room holds and processes all the GameItems and
draws them to the screen.

I also created and implemented classes for the Tower and each Zombie
type. Finally, there is a working update loop in place.