Added Player and PlayerUI
Player tracks the players hp and cash, which can eventually be used as
zombies leave the screen or are killed, and for making new towers. These
values are shown on screen through UI, and there is an Observer system
set up so that when the players hp or cash change, it can send a message
to its observer, the ui to change the values displayed. Still to do is
the pathfinding algorithm and unit tests.